% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.r
\docType{data}
\name{africa_healthsites_WHO}
\alias{africa_healthsites_WHO}
\alias{sf_who_sites}
\title{africa healthsite points from WHO}
\format{\code{sf}}
\source{
\url{https://www.who.int/malaria/areas/surveillance/who-cds-gmp-2019-01-eng.xlsx}
}
\usage{
sf_who_sites
}
\description{
africa healthsite points from WHO
}
\section{Slots}{

\describe{
\item{\code{data}}{healthsite points}
}}

\keyword{datasets}
