% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.r
\docType{data}
\name{african_country_names}
\alias{african_country_names}
\alias{afcountries}
\title{african country names and iso 3 letter country codes}
\format{\code{dataframe}}
\usage{
afcountries
}
\description{
african country names and iso 3 letter country codes
}
\section{Slots}{

\describe{
\item{\code{data}}{dataframe}
}}

\keyword{datasets}
