% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_rhealthsites.r
\name{check_rhealthsites}
\alias{check_rhealthsites}
\title{Check whether to install rhealthsites and install if necessary}
\usage{
check_rhealthsites()
}
\description{
If the rhealthsites package is not installed, install it from GitLab using
remotes.
}
