% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/country2iso.r
\name{country2iso}
\alias{country2iso}
\title{conversion from country names to iso3c code}
\usage{
country2iso(country)
}
\arguments{
\item{country}{a character vector of country names}
}
\value{
character vector of iso3c codes
}
\description{
#todo vectorise
}
\examples{

iso3c <- country2iso("nigeria")

}
