% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_rhealthsites.r
\name{install_rhealthsites}
\alias{install_rhealthsites}
\title{Install the rhealthsites package after checking with the user}
\usage{
install_rhealthsites()
}
\description{
Install the rhealthsites package after checking with the user
}
