% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/country2iso.r
\name{iso2country}
\alias{iso2country}
\title{conversion from iso3c code to country names}
\usage{
iso2country(iso3c)
}
\arguments{
\item{iso3c}{a character vector of country codes}
}
\value{
character vector of country names
}
\description{
#todo vectorise
}
\examples{

name <- iso2country("nga")

}
