% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_points.r
\name{merge_points}
\alias{merge_points}
\title{merge healthsite points from different sources for a country}
\usage{
merge_points(
  country,
  datasources = c("healthsites", "who"),
  hs_amenity = "all",
  dist_same_m = 100
)
}
\arguments{
\item{country}{a character vector of country names or iso3c character codes.}

\item{datasources}{vector of 2 datasources from 'healthsites' predownloaded, 'who', 'healthsites_live' needs API, 'hdx' not working yet}

\item{hs_amenity}{filter healthsites data by amenity. 'all', 'clinic', 'dentist', 'doctors', 'pharmacy', 'hospital'}

\item{dist_same_m}{distance below which a site from source 1 and 2 is considered same}
}
\value{
\code{sf}
}
\description{
IN DEVELOPMENT
aim is to identify and remove duplicates
first version just tests distance, between points in layer1 and 2
BUT problem maybe that there are points in each layer that are duplicates too
may want to move this to separate package later
}
\examples{

#merge_points("nigeria", datasources=c('who', 'healthsites'), plot='mapview')
#merge_points(c('malawi','zambia'))

}
