% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nameof_.r
\name{nameof_labcol}
\alias{nameof_labcol}
\title{name of label column according to datasource}
\usage{
nameof_labcol(datasource)
}
\arguments{
\item{datasource}{vector of 2 datasources from 'healthsites' predownloaded, 'who', 'healthsites_live' needs API, 'hdx' not working yet}
}
\value{
character column name
}
\description{
name of label column according to datasource
}
\examples{

nameof_labcol('who')

}
