% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nameof_.r
\name{nameof_zcol}
\alias{nameof_zcol}
\title{name of z column according to datasource}
\usage{
nameof_zcol(datasource, type_column)
}
\arguments{
\item{datasource}{'healthsites' predownloaded, 'who', 'healthsites_live' needs API, 'hdx' not working yet}

\item{type_column}{just for user provided files which column has information on type of site, default : 'Facility Type'}
}
\value{
character column name
}
\description{
name of z column according to datasource
}
\examples{

nameof_zcol('who')

}
