% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/national_list_avail.r
\name{national_list_avail}
\alias{national_list_avail}
\title{availability of a national master facility list by country}
\usage{
national_list_avail(country = "all")
}
\arguments{
\item{country}{'all' (default) or a single country name or iso3c character code.}
}
\value{
dataframe or single value
}
\description{
returns data availability dataframe for 'all' countries or 1 row for a single country
TODO add calculation of stats on numbers of countries satisfying criteria
}
\examples{
national_list_avail('Togo')

}
