% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/national_list_url.r
\name{national_list_url}
\alias{national_list_url}
\title{data url for national master facility lists by country}
\usage{
national_list_url(country = "all")
}
\arguments{
\item{country}{'all' (default) or a single country name or iso3c character code.}
}
\value{
dataframe or single value
}
\description{
returns urls for 'all' available countries or url for a single country
}
\examples{
national_list_url('South Sudan')

#example of reading in data direct from a url and mapping
#will only work for countries where "machine_readable" is TRUE
#dfgha <- read.csv(national_list_url("Ghana"))
#sfgha <- sf::st_as_sf(dfgha, coords=c("Longitude","Latitude"), crs=4326, na.fail=FALSE)
#mapview::mapview(sfgha, zcol='Type')
#TODO fix why this fails
#compare_hs_sources("ghana",datasources=c("who",sfgha))
#in compare_hs_sources country=ghana admin_level= admin_names=
#Error in afrihealthsites(country, datasource = datasources[[2]], plot = FALSE,  :
#                            object 'sfcountry' not found

#to return a dataframe with all countries that have urls
national_list_url()

}
