% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/near_points.r
\name{near_points}
\alias{near_points}
\title{find near points for a country initially in one datasource}
\usage{
near_points(
  country,
  datasources = c("healthsites", "who"),
  hs_amenity = c("clinic", "doctors", "pharmacy", "hospital"),
  dist_same_m = 50,
  toreturn = "summary"
)
}
\arguments{
\item{country}{a character vector of country names or iso3c character codes.}

\item{datasources}{vector of 2 datasources from 'healthsites' predownloaded, 'who', 'healthsites_live' needs API, 'hdx' not working yet}

\item{hs_amenity}{filter healthsites data by amenity. 'all', 'clinic', 'dentist', 'doctors', 'pharmacy', 'hospital'
to exclude dentist hs_amenity=c('clinic', 'doctors', 'pharmacy', 'hospital')}

\item{dist_same_m}{distance below which a site from source 1 and 2 is considered same}

\item{toreturn}{whether to return 'summary' or the merged 'sf' object}
}
\value{
\code{sf}
}
\description{
IN DEVELOPMENT
aim is to help identify and remove duplicates
}
\examples{

#near_points("burundi", datasources=c('who', 'healthsites'), plot='mapview')

}
