% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.r
\docType{data}
\name{south_sudan_moh}
\alias{south_sudan_moh}
\alias{sfssd}
\title{south sudan health facility points from moh}
\format{
\code{sf}

An object of class \code{sf} (inherits from \code{data.frame}) with 2889 rows and 18 columns.
}
\source{
\url{https://www.southsudanhealth.info/facility/fac.php?list&s=0&p=0&ps=2889}
}
\usage{
sfssd
}
\description{
south sudan health facility points from moh
}
\section{Slots}{

\describe{
\item{\code{data}}{health facility points}
}}

\keyword{datasets}
