% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_error.r
\name{test_error}
\alias{test_error}
\title{temporary attempt to create reproducible example for sf error 'does not point to geometry column'}
\usage{
test_error(plot = "sf")
}
\arguments{
\item{plot}{option to display map 'mapview' for interactive, 'sf' for static}
}
\value{
\code{sf}
}
\description{
temporary attempt to create reproducible example for sf error 'does not point to geometry column'
}
\examples{


tst1 <- test_error(plot='mapview')


}
