% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tiers4.r
\name{tiers4}
\alias{tiers4}
\title{takes a vector of facility types and classes them into 4 tiers following Falchetta type definition}
\usage{
tiers4(
  types,
  tier1 = NULL,
  tier2 = NULL,
  tier3 = NULL,
  tier4 = NULL,
  nametier0 = "Tier0 unknown",
  nametier1 = "Tier1 health post",
  nametier2 = "Tier2 health centre",
  nametier3 = "Tier3 provincial hospital",
  nametier4 = "Tier4 central hospital",
  to_return = "Tier"
)
}
\arguments{
\item{types}{facility types, one per health facility as a vector}

\item{tier1}{types to be classed as tier1}

\item{tier2}{types to be classed as tier2}

\item{tier3}{types to be classed as tier3}

\item{tier4}{types to be classed as tier4}

\item{nametier0}{name for tier0 i.e. when it isn't classed into other tiers}

\item{nametier1}{name for tier1}

\item{nametier2}{name for tier2}

\item{nametier3}{name for tier3}

\item{nametier4}{name for tier4}

\item{to_return}{"Tier" or "Tier_name" potentially others later}
}
\value{
vector of numeric values or strings representing tiers
}
\description{
to reclass facility types into a 4 tier classification, any types not passed will be classed as 0
}
\examples{

tiers4(c("a","b","c"), tier1=c("a","b"), tier2="c", to_return="Tier_name")

}
