% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.r
\docType{data}
\name{who_type_lookup}
\alias{who_type_lookup}
\title{lookup table to convert 173 WHO facility types to 9 broad categories}
\format{
\code{dataframe}

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 318 rows and 3 columns.
}
\source{
\url{https://link.springer.com/article/10.1186/s12916-019-1459-6}
}
\usage{
who_type_lookup
}
\description{
lookup table to convert 173 WHO facility types to 9 broad categories
}
\section{Slots}{

\describe{
\item{\code{data}}{who type lookup}
}}

\keyword{datasets}
