% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.r
\docType{data}
\name{africa_healthsites}
\alias{africa_healthsites}
\alias{sf_healthsites_af}
\title{africa healthsite points from healthsites.io}
\format{\code{sf}}
\source{
\url{https://www.healthsites.io}
}
\usage{
sf_healthsites_af
}
\description{
africa healthsite points from healthsites.io
}
\section{Slots}{

\describe{
\item{\code{data}}{healthsite points}
}}

\keyword{datasets}
