% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afrihealthsites.r
\name{afrihealthsites}
\alias{afrihealthsites}
\title{Get healthsite points for a country}
\usage{
afrihealthsites(country, datasource = "healthsites", plot = "mapview")
}
\arguments{
\item{country}{a character vector of country names or iso3c character codes.}

\item{datasource}{data source, initial default 'who'}

\item{plot}{option to display map 'mapview' for interactive, 'sf' for static}
}
\value{
\code{sf}
}
\description{
returns healthsite locations for specified countries and optionally plots map
}
\examples{

sfnga <- afrihealthsites("nigeria")

}
