% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afrihealthsites-package.r, R/afrihealthsites.r
\docType{package}
\name{afrihealthsites}
\alias{afrihealthsites}
\title{afrihealthsites : healthsite locations for Africa}
\usage{
afrihealthsites(
  country,
  datasource = "healthsites",
  plot = "mapview",
  hs_amenity = "all",
  who_type = "all",
  type_filter = "all",
  returnclass = "sf",
  type_column = "Facility Type",
  label_column = "Facility Name",
  lonlat_columns = c("Longitude", "Latitude"),
  admin_level = NULL,
  admin_names = NULL
)
}
\arguments{
\item{country}{a character vector of country names or iso3c character codes.}

\item{datasource}{data source, 'healthsites' predownloaded, 'who', 'healthsites_live' needs API, 'hdx' not working yet}

\item{plot}{option to display map 'mapview' for interactive, 'sf' for static}

\item{hs_amenity}{filter healthsites data by amenity. 'all', 'clinic', 'dentist', 'doctors', 'pharmacy', 'hospital'
to exclude dentist hs_amenity=c('clinic', 'doctors', 'pharmacy', 'hospital')}

\item{who_type}{filter by Facility type}

\item{type_filter}{filter by facility type (from the type_column, or internally defined for healthsites & who)}

\item{returnclass}{'sf' or 'dataframe', currently 'dataframe' only offered for WHO so that can have points with no coords}

\item{type_column}{for user provided files which column has information on type of site, default : 'Facility Type'}

\item{label_column}{for user provided files which column has information on name of site, default : 'Facility Name'}

\item{lonlat_columns}{for user provided files which columns contain longitude, latitude. option of NULL if no coords}

\item{admin_level}{what admin level to filter regions from  NULL if no filtering}

\item{admin_names}{names of admin regions to filter NULL if no filter}
}
\value{
\code{sf}
}
\description{
locations from WHO and healthsites.io. Part of afrimapr project.

returns healthsite locations for specified countries and optionally plots map
}
\examples{

sfnga <- afrihealthsites("nigeria", datasource='who', plot='sf')

afrihealthsites('chad', datasource='who', plot='sf')
afrihealthsites('chad', datasource='healthsites', plot='sf')

sfnga <- afrihealthsites("nigeria", plot='mapview')

#to return raw dataframe for WHO data including any rows with no coordinates
dfzaf <- afrihealthsites("south africa", datasource='who', plot=FALSE, returnclass='dataframe')
#note that ISO 3 letter codes and country names with upper case letters can also be used
#afrihealthsites("ZAF")
#afrihealthsites("South Africa")


#filter healthsites data by amenity type
afrihealthsites('chad',datasource = 'healthsites', hs_amenity=c('clinic','hospital'))
#filter who data by Facility type
afrihealthsites('chad',datasource = 'who',who_type=c('Regional hospital','Health Centre'))

#filter by admin regions
afrihealthsites('togo', admin_level=1, admin_names=c('Maritime Region', 'Centrale Region'))

}
\seealso{
\code{\link{afrihealthsites}} \code{\link{compare_hs_sources}}
}
