% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afrihealthsites-package.r, R/afrihealthsites.r
\docType{package}
\name{afrihealthsites}
\alias{afrihealthsites}
\title{afrihealthsites : healthsite locations for Africa}
\usage{
afrihealthsites(
  country,
  datasource = "healthsites",
  plot = "mapview",
  hs_amenity = "all",
  returnclass = "sf"
)
}
\arguments{
\item{country}{a character vector of country names or iso3c character codes.}

\item{datasource}{data source, 'healthsites' predownloaded, 'who', 'healthsites_live' needs API, 'hdx' not working yet}

\item{plot}{option to display map 'mapview' for interactive, 'sf' for static}

\item{hs_amenity}{filter healthsites data by amenity. 'all', 'clinic', 'dentist', 'doctors', 'pharmacy', 'hospital'
to exclude dentist hs_amenity=c('clinic', 'doctors', 'pharmacy', 'hospital')}

\item{returnclass}{'sf' or 'dataframe', currently 'dataframe' only offered for WHO so that can have points with no coords}
}
\value{
\code{sf}
}
\description{
locations from WHO and healthsites.io. Part of afrimapr project.

returns healthsite locations for specified countries and optionally plots map
}
\examples{

sfnga <- afrihealthsites("nigeria", datasource='who', plot='sf')

afrihealthsites('chad', datasource='who', plot='sf')
afrihealthsites('chad', datasource='healthsites', plot='sf')

sfnga <- afrihealthsites("nigeria", plot='mapview')

#to return raw dataframe for WHO data including any rows with no coordinates
dfzaf <- afrihealthsites("south africa", plot=FALSE, returnclass='dataframe')


}
\seealso{
\code{\link{afrihealthsites}} \code{\link{compare_hs_sources}}
}
