% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_hs_sources.r
\name{compare_hs_sources}
\alias{compare_hs_sources}
\title{compare healthsite points from different sources for a country}
\usage{
compare_hs_sources(
  country,
  datasources = c("healthsites", "who"),
  plot = "mapview",
  plotshow = TRUE,
  plotcex = c(6, 3),
  col.regions = list(RColorBrewer::brewer.pal(5, "YlGn"), RColorBrewer::brewer.pal(5,
    "BuPu")),
  plotlegend = TRUE
)
}
\arguments{
\item{country}{a character vector of country names or iso3c character codes.}

\item{datasources}{vector of 2 datasources from 'healthsites' predownloaded, 'who', 'healthsites_live' needs API, 'hdx' not working yet}

\item{plot}{option to display map 'mapview' for interactive, 'sf' for static}

\item{plotshow}{whether to show the plot, otherwiser just return plot object}

\item{plotcex}{sizes of symbols for each source default=c(6,3), helps view symbol overlap}

\item{col.regions}{list of two colour palettes to pass to mapview}

\item{plotlegend}{whether to add legend to mapview plot}
}
\value{
\code{sf}
}
\description{
main aim to plot map comparing different sources
}
\examples{

#compare_hs_sources("nigeria", datasources=c('who', 'healthsites'), plot='mapview')
#compare_hs_sources(c('malawi','zambia'))

}
