% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_hs_sources.r
\name{compare_hs_sources}
\alias{compare_hs_sources}
\title{compare healthsite points from different sources for a country}
\usage{
compare_hs_sources(
  country,
  datasources = c("healthsites", "who"),
  plot = "mapview",
  plotshow = TRUE,
  plotcex = c(6, 4),
  col.regions = list(RColorBrewer::brewer.pal(5, "YlGn"), RColorBrewer::brewer.pal(5,
    "BuPu")),
  alpha = c(0.1, 0.1),
  alpha.regions = c(0.7, 0.7),
  layer.names = NULL,
  plotlegend = TRUE,
  hs_amenity = "all",
  who_type = "all",
  canvas = TRUE,
  plotlabels1 = FALSE,
  plotlabels2 = FALSE,
  map.types = c("CartoDB.Positron", "OpenStreetMap.HOT"),
  type_column = "Facility Type",
  label_column = "Facility Name",
  lonlat_columns = c("Longitude", "Latitude")
)
}
\arguments{
\item{country}{a character vector of country names or iso3c character codes.}

\item{datasources}{vector of 2 datasources from 'healthsites' predownloaded, 'who', 'healthsites_live' needs API, 'hdx' not working yet}

\item{plot}{option to display map 'mapview' for interactive, 'sf' for static}

\item{plotshow}{whether to show the plot, otherwiser just return plot object}

\item{plotcex}{sizes of symbols for each source default=c(6,3), helps view symbol overlap}

\item{col.regions}{list of two colour palettes to pass to mapview}

\item{alpha}{list of two alphas to pass to mapview - low keeps borders light}

\item{alpha.regions}{list of two alpha.regions to pass to mapview}

\item{layer.names}{allow mapview layer.names to be set c('a','b')}

\item{plotlegend}{whether to add legend to mapview plot}

\item{hs_amenity}{filter healthsites data by amenity. 'all', 'clinic', 'dentist', 'doctors', 'pharmacy', 'hospital'}

\item{who_type}{filter by Facility type}

\item{canvas}{mapview plotting option, TRUE by default for better performance with larger data}

\item{plotlabels1}{whether to add static labels for source1}

\item{plotlabels2}{whether to add static labels for source2}

\item{map.types}{optional specification of background map tiles for mapview, default c('CartoDB.Positron','OpenStreetMap.HOT')}

\item{type_column}{just for user provided files which column has information on type of site, default : 'Facility Type'}

\item{label_column}{just for user provided files which column has information on name of site, default : 'Facility Name'}

\item{lonlat_columns}{for user provided files which columns contain longitude, latitude. option of NULL if no coords}
}
\value{
\code{sf}
}
\description{
main aim to plot map comparing different sources
}
\examples{

#compare_hs_sources("nigeria", datasources=c('who', 'healthsites'), plot='mapview')
#compare_hs_sources(c('malawi','zambia'))

}
