% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facility_types.r
\name{facility_types}
\alias{facility_types}
\title{get & plot freq of facility types for a country}
\usage{
facility_types(
  country,
  datasource = "healthsites",
  type_filter = "all",
  plot = TRUE
)
}
\arguments{
\item{country}{a character vector of country names or iso3c character codes.}

\item{datasource}{data source, 'healthsites' predownloaded, 'who', 'healthsites_live' needs API, 'hdx' not working yet}

\item{type_filter}{filter healthsites data by amenity. 'all', 'clinic', 'dentist', 'doctors', 'pharmacy', 'hospital'
to exclude dentist hs_amenity=c('clinic', 'doctors', 'pharmacy', 'hospital')}

\item{plot}{whether to display plot}
}
\value{
\code{sf}
}
\description{
get & plot freq of facility types for a country
}
\examples{

sfnga <- facility_types("nigeria", datasource='who', plot='sf')

facility_types('chad', datasource='who', plot='sf')
facility_types('chad', datasource='healthsites', plot='sf')


#filter healthsites data by amenity type
facility_types('chad',datasource = 'healthsites', hs_amenity=c('clinic','hospital'))
#filter who data by Facility type
facility_types('chad',datasource = 'who',who_type=c('Regional hospital','Health Centre'))


}
