% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nameof_.r
\name{nameof_labcol}
\alias{nameof_labcol}
\title{name of label column according to datasource}
\usage{
nameof_labcol(datasource, label_column)
}
\arguments{
\item{datasource}{vector of 2 datasources from 'healthsites' predownloaded, 'who', 'healthsites_live' needs API, 'hdx' not working yet}

\item{label_column}{just for user provided files which column has information on name of site, default : 'Facility Name'}
}
\value{
character column name
}
\description{
name of label column according to datasource
}
\examples{

nameof_labcol('who')

}
