% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/national_facility_list.r
\name{national_list_avail}
\alias{national_list_avail}
\title{availability of a national master facility list by country}
\usage{
national_list_avail(country)
}
\arguments{
\item{country}{a character vector of country names or iso3c character codes.}
}
\value{
dataframe or single value
}
\description{
initially return whole row for countries
later add a function partic to access 'url_data'
}
